Disable recording admin logins
I guess this line of code will disable recording when an admin access the client's account (only works if the admin logins using the "Login as Client" button from client summary page. i.e: if admin logins directly from clientarea using email and password of the client, this will execute the rest of the hook).